Refactor/#66 api, repository 클린아키텍처 구조로 일부 변경
- 프로필 페이지 라우트 변경 :
