+++ src/app/user/profile/page.tsx
@@ -1,0 +1,7 @@
+import ProfilePage from "@/components/profile/ProfilePage";
+
+const profilePage = () => {
+    return <ProfilePage />;
+};
+
+export default profilePage;
